backend/pyproject.toml
[tool.poetry]
name = "SSC0952"
version = "0.1.0"
description = ""
authors = ["Allan Garcia"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.0.3"
psycopg2-binary = "^2.9.9"
paho-mqtt = "^1.6.1"
google = "^3.0.0"
protobuf = "^4.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
